Move `chia.consensus.fee_estimate_store` to `chia.full_node` (#19705)
The `chia.protocols` module no longer depends upon `chia.full_node`
