+++ chia/_tests/core/mempool/test_mempool_fee_estimator.py
@@ -9,11 +9,11 @@ from chia._tests.core.consensus.test_pot_iterations import test_constants
 from chia._tests.util.db_connection import DBConnection
 from chia.full_node.bitcoin_fee_estimator import BitcoinFeeEstimator
 from chia.full_node.coin_store import CoinStore
+from chia.full_node.fee_estimate_store import FeeStore
 from chia.full_node.fee_estimation import MempoolItemInfo
 from chia.full_node.fee_estimator import SmartFeeEstimator
 from chia.full_node.fee_tracker import FeeTracker
 from chia.full_node.mempool_manager import MempoolManager
-from chia.protocols.fee_estimate_store import FeeStore
 
 
 @pytest.mark.anyio
