+++ chia/full_node/bitcoin_fee_estimator.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 from chia_rs.sized_ints import uint32, uint64
 
+from chia.full_node.fee_estimate_store import FeeStore
 from chia.full_node.fee_estimation import EmptyFeeMempoolInfo, FeeBlockInfo, FeeMempoolInfo, MempoolItemInfo
 from chia.full_node.fee_estimator import SmartFeeEstimator
 from chia.full_node.fee_estimator_interface import FeeEstimatorInterface
 from chia.full_node.fee_tracker import FeeTracker
-from chia.protocols.fee_estimate_store import FeeStore
 from chia.types.clvm_cost import CLVMCost
 from chia.types.fee_rate import FeeRateV2
